Split retrieving of licenses to its own module so it can be imported by external scripts
